zsh: make grep case insensitive by default
Case insensitive use-cases for grep outnumber
the case-sensitive ones significantly, at least for me.

This only applies to regular grep invocation.
I do care for case when I `fgrep` or `egrep`.

Sadly, there is no --preserve-case option for grep,
so this means that if I ever want to grep
with case-sensitivity I'll now have to do

$ command grep ...

The tradeoff is worth it, at least right now.