zsh & git: add support for cyan git status in prompt
While green, yellow, and red mean "in sync with something", "not
synched yet", and "out of sync", respectively, cyan means "syncing is
not applicable." This means that no remote exists in case of the
branch component, or that this is a bare repository in case of the sha
component. It doesn't make sense to compare local status with remote
branches or local worktree in either case, so cyan is introduced to
signal this nonapplicability, which was previously simply shown in
green.