i3: bind functionality to some XF86 multimedia keys
These packages are required for the new functionality:
  - light
  - mpc

The XF86MonBrightness{Up,Down} bindings obsolete the `brightness`
shell function.

Note that systemd-backlight@.service(8) always resets backlight to to
full brightness when rotating the screen via xrandr if the current
brightness is less than its minimum permissible brightness, which is
easy to accidentally trigger via the keybinding. As a workaround, you
can use light(1)'s -N switch to set a soft lower limit on the
brightness to which the XF86MonBrightness binding can decrement that's
above systemd-backlight@.service(8)'s lower threshold.