zsh: add wrapper fn for cd
This is supposed to behave like cd in most cases,
but be selectively smarter when the context allows it,
particularly when it would fail instead otherwise.

If we're jumping to the dirs containing files when cd'ing
to the files already, we might as well do so for executables
on PATH as well, which has the added benefit of synergizing
well with NixOS's nix store.